Added sign in/out support for Google accounts.
Added:
- Sign in page
-- Contains 1 button for signing in with Google account.
- Account page
-- Contians 1 button for signing out of account.

Account section (the right-hand section) of header now uses a template. Previous the same code was used twice. Now they both have only 1 line that points to a single template.

Account section of header updates based on whether user is signed in.
If signed out, it shows 'Sign In' and linked to sign in page.
If signed in, it shows 'Account' and links to account page.

Account page currently shows the users SocialUser object in the console. This is for testing purposes and must be removed before the application is released.