Now handles isAdmin status when authenticating with backend.
API now returns whether user is an admin when authenticating.
SignInService now stores this as an observable. The header subscribes to this.

When user is signed in as an admin the account section of the header now shows 'ADMIN' as leftmost item.
Current does nothing. Just shows user is an admin.