Authorization header now auto-injected into requests.
API service now holds auth object when user signs in.

API service now implements get, post, delete http methods.
These are used instead of the regular methods.
Uses the regular methods, but automatically injects the user JWT as HTTP_IDTOKEN if it is set, allowing for the user to do requests that require authorization with the API.