Updated topics list to support delete.
Topic list reworked.
Now included 'View Topic' button. This button must be pressed to navigate to the topics area.

Is signed in as an admin, 'Edit', and 'Delete' buttons are shown for each topic item. The delete button will delete the topic.
The delete button validates that the user is in fact an admin before attempting to delete. (and the backend also checks)