Updated subject creator. Now uses ngx-quill text editor.
Was experiencing problems with tinymce and the 'hacky' solution required to get it to work.

Now uses ngx-quill, which works without any 'hacky' stuff going on.

Page now displays current length of encoded body, updated when the text is changed. Hopefully helps user, since the encoded body has to be 65,535 chars or less in length.