Messages sent to / from lists now show user name at top.
On messages sent to list, the name of the user the messages have been sent to is shown. Their name is no longer shown on the individual messages.

On messages from list, the name of the user who sent the messages is shown. Their name is no longer shown on the individual messages.