Added search functionality to user list.
Can now search users on the user list via an input.
Search is done when input value is changed.
Search is delayed by 1000ms. Each time the user inputs a character into the input, the delay is reset back to 1000ms, acting like a throttle.

Also updated the getUsers method.
Now if the user is searching, the fetched records are filtered and added to the end of the search list.

If the search list length is less than count, and the user is currently searching, and the end of content has not been reached, getUsers will be repeatedly called, with each call incrementing the offset. This is because a flaw with infinite scroll, where if too few users are listed for the page scroll bar to be shown, it is impossible to get more users, as it is impossible to trigger the scroll event for the infinite scroll.