Added basic handling of banned users when they attempt to sign in.
Previously when banned users signed in, they could do so successfully, however they would fail to authorize with the api.

This put them in a sort of limbo state, where the frontend considered them signed in, but was unable to make any api calls that required authorizations, which caused any functionality requiring authorization to break.

Current fix is kind of a hack and should be improved later.
If server returns 401 when the user attempts to authorise initially (with their account id_token), they will be shown and alert and automatically signed out.