Added confirmation modal for deleting subjects.
Added confirmation modal for deleting subjects.
Replaces confirm() method previously used.

Stored index of subject in deleteSubjectIndex variable.
This determines which subject is deleted.
This is set to null after the subject is deleted.
This value is maintained if the modal is cancelled, because jQuery does not play well with angular. Should not be a problem however, as the value is updated when a new delete modal is shown.

Other:
- Content description is now has height of 5.8em. Previous only the max-height was 5.8em. This is for consistency. Now all content cards have same height regardless of description size.