Updated lessons/tests lists to use grid style.
Updated lessons/tests lists to use grid style, as well as use the new style of cards (with the buttons in the footer).

They both also use a modal for delete confirmations now.

Updated both, as well as the subjects and topics lists to have a better styled 'no more content message'. (The message at the bottom of the list, when all piece of content have been fetched).
They are now centered, have center-aligned text, use bootstrap dark color-scheme.
They have width of 30em, with max-width of 100%.