Added beta group list (see details)
Early version of group list added.
Displays groups that the current user is a member of.
If user is member of no groups, a message is shown stating this.

Uses grid styling.

Each card has an image on it for the groups imageUrl. If none is given, it instead uses a 256x256 placeholder image from https://placehold.it/.
Does not current valid the image.
Does not current style the image correctly. E.g. using a 512x512 image completely messages up the styling of the grid, with the column the image appears in becoming disproportionately wide.

Opening the group list performs validation checks:
- User must be signed in.
The user admin status is gotten on page load. Not currently used. Planned to be used to determine with an 'all groups' button should be shown. This will take the user to a groups list displaying all groups.

Page has create button, and view buttons on each card, however, none are usable.

Page does not yet have any kind of breadcrumbs or back button.