Updated style of test questions list.
Test questions list now uses grid styling of most other lists.
Aka shows 1/2/3 columns depending on screen width.

Cards no longer have headers.
Top of cards contains the questions image, if set. If not set, the space in filled with a div, with a light gray background, and some text informing the user that the question has no image.
The image/div is set to be 20em tall. This can cause some warping of the image if it is overly wide.
The user will be able to see the image displayed correctly via the test question home page once further implemented. This pages will be extremely basic, just showing the question, answer, and image (with no size limits).