Reworked news post area.
Removed news post list components that was shared due to it being unwieldly and the resulting code-reuse caused by removing it being miniscule.

News posts now displayed differently.
The contain a grey badge for the last modified date, at the top, 0.2em from the card border.
They contain the title and below below, in similar style to the content grid.
They have a footer, only visible to admins that contains an edit (current unusable) and delete button.
Delete button opens confirmation modal.

Updated subject home to show newest 3 news posts.
Only 3 are shown. It is determined by a variable in the typescript for the home component. This is no way to load more.
The admin controls are not shown on the news posts on in the subject home area.

Discovered bug with delete for contents lists, was fixed for news post list, but must be checked for on other content lists.

When a post is deleted, the offset is not decremented by 1. This means that 1 record will be skipped when loading more records after deleting a record.

---
Updated some styling. This commit was meant to be separate but I didn't notice I didn't commit the previous update...

Moved the CSS for the navigation bar (header), and animated app loading icon, from styles.css to the respective component.

Updated user list. View button is now in button left corner. Dropdown no longer contains view button or divider. Dropdown is now only shown to admins.
The badge for a user (if they have one) is now shown in the bottom left of the card body, with a margin of 0.2em.
The user list search has been moved into a card in an attempt to make it look nicer.