Added create group page. Added breadcrumb to group list and home.
Added create group page.
Has similar style to other creates pages.
Has inputs for:
- name
- description
- image url

Has a preview based on this values, that updates as the values change.
The image url is checked to ensure it is valid.
If it is not valid, the user will be unable to submit, and a message will be shown below the image url input.

Added breadcrumb navigation to group list and home pages.