Added edit group page. Group descriptions now support Markdown.
Added edit group page. Works as expected, like group create, but pre-populates fields with existing values, calls different api method, ensures only changed values are sent.

Fixed bug on group creator. Previously it redirect to group list when the group was created. It now correctly redirects to group home for the create group.

The descriptions of groups on both the group list and group home are now rendered as Markdown.