Added support for subject admins.
Subject home now shows list of admins associated with the subject.
Is in a table.
Can chat with any user via a button next to each.

If you are an admin, there is an option to associate or dissassociate with the subject. The button that is shown depends on whether you are currently associated with the subject.