Appropriate content lists now show admins if a piece of content has been hidden or auto-hidden.
Hidden and auto-hidden support finally being implemented.
Involved: Subjects, topics, lessons, tests. Both lists and home areas.

If the content has its hidden attribute set, it will be given a badge stating so.
If the content has no child elements (topics for subjects, lessons and tests for topics, test questions for tests), it will be given a badge stating so.
Only 1 of the 2 badges can be shown for a piece of content, with the regular hidden badge taking precedence.

The badge is shown in the bottom right of the card body that the piece of content is displayed on.