Updated system so that it can successfully build into production mode.
System was previous unable to build into production mode, as the typescript files for components had some private properties being accessed from the html of the component.

This compiles fine in dev mode, however will not compile in prod mode.

This resulted in many variables being changed to public.
This should not be a problem as private/public in typescript does not technically matter anyway. The result JavaScript code will not include the private/public accessibility.

The outputted production code was tested on localhost and worked successfully, however for an unknown reason, the user's login state would not be remember, e.g they could sign into a social media account successfully, however after refreshing the page they would be signed out and would not be automatically signed back in automaitcally.