Added global debug params [see inside...]
Added ```debug``` parameter to environment variable.

Contains 2 attributes:
- ```showSocialMediaSignInObjects```
  - When set, the social user object returned by ngx-social-login when the user signs in with their social media account is logged to the console.
- ```showAPIAuthorizationObjects```
  - When set, the authorization data returned by the API when the user successfully authentications with the API is logged to the console.

Both should be set to false for production.
Both exists for testing reasons. Both so that the sign in with Google/Facebook and authenticate with the API can be tested.
And because testing of the API requires social media id tokens for testing the sign in methods. This allows access to the id tokens.