Added dynamic line chart to user test lists page.
User tests list page has been updated to include a dynamic line chart.
The chart plots the users test scores (percentages).
Some points are labelled by the date that they were completed. The labels are on the x axis.
A max of 6 labels can be shown at once, with additionally labels not being shown. The labels are automatically distributed so that labels each appear every ~1/6 of the charts width.

Each point can be hovered over. This will shown a tooltip showing both the score and the label (completion date). This ensures that points label can be seen even if it is hidden on the x axis.