[Testing] - Updated group list and home so that admins can better view groups they are not a member of.
GROUPS LIST:
Updated group list to have a button that only admins can see.
Toggles list to show all groups, instead of just 'my groups'. Added title to page that changes between 'My Groups' and 'All Groups' approrpriately.
When viewing all groups list, the bottom right dropdown is not shown, instead only an edit button is shown.
This is because the logic for handling whether the admin is a member/non-member of each group would be difficult to implement.
Updated leave button on group list. Previously did nothing. Now it shows a confirmation modal. If confirmed, the user leaves the group, and the list is updated.

GROUP HOME:
Admins can now better view the group home of groups they are not a member of. Previous the leave button would cause an error if the admin was not a member of the group.
Now if the user is an admin, an additional api call is made to check whether the user is in the group.
If they are in the group, the leave button is shown.
If they are not in the group, a join button is shown. Admins may forcably join groups.

Added 2 methods to group service, one for getting all groups, one for getting whether the current user is a member of a specific group.