[Testing] - App now checks current route on navigation end. Clears subject if route is /subjects.
The app now checks the current route on navigation end, via the router events.
If the event is a NavigationEnd, and the events urlAfterRedirects value is '/subjects', the subject is cleared on the subject service, resulting in the navigation bar automatically updating to how it should be when not inside a subject.

This solves the issue where the 'Subjects' link of the breadcrumb navigation on most pages would not update the navigation bar when used.

The solution was either this, or update every single page so clear the subject when the 'subjects' link is used.
This seems like a far cleaner and more maintainable solution...