[Testing] - Added some validation to a couple editors to stop issues being caused by purely whitespace inputs.
Can no longer set name to whitespace, e.g. '     ', on the following editors:
- Subject editor
- Subject news editor
- Group editor.

Image url is now trimmed before being sent to API. Done for:
- Group creator/editor

This was a problem, as if the image url was '', the placeholder is shown, however, if it is purely whitespace, e.g. '     ', it would instead display the normal HTML image missing stuff.