[Testing] - Fixed validation of inputs on user test attempt. Updated tables to have overflow-x: auto, so that they behave better on mobile / with very long text.
Updated validation for user test attempt to user correct values.
Namely:
title - maxlength: 30
each question input - maxlength: 255

Also updated how default title works. Previously it was setting the name to the current time, even if the user entered a custom title. Now it uses the custom title if set, otherwise uses current time.

Updated all tables used on pages in the system.
Now has width:100% and overflow-x: auto.
Results in tables becoming scrollable on x axis if screen width is too small.
The table scrolls, not the whole pages. Previously the table contents was extend out of the table and offscreen. To view it you had to scroll the page in general, and you could see the table content extented out of the table.