Updated Markdown sanitization to work considerably better.
Updated Markdown sanitization.
No longer users the sanitization built into ngx-markdown. This is because it was not the best. It did not block all unsafe HTML and blocked some useful safe HTML, such as p, div, and center tags, as well as blocked style attributes on tags such as p/div.

The sanitization is now done by DOMPurify, which fixes both of the previously mentioned issues.

This resulted in a method named sanitizeMarkdown being added to the UtilService.
It also meant that many pages, such as the subject, topic, lesson, test, and group list/home,create, and edit pages being updated to call this method before displaying the Markdown.