Updated lesson creator/editor to use a rich text editor editor again...
Update lesson creator/editor to use a rich text editor (ngx-quill, which is based on quilljs).

Fixed previous problem with alignment. Turns out quill outputs the HTML with classes for styling. I changed some config values to make it use style attributes.

The output of the quill editors is passed through the sanitizeMarkdown method to ensure it does not contain anything dodge.

The editors allow for significantly easier creation/editing of lessons, as they are much easier and are like a normal text editor. They also allow alignment, which is extremely awkward to use with the regular Markdown content descriptions.

The editors help to differentiate lessons from other content, previously you could effectively have a lesson in a description, such a subject description. This is still somewhat possible, however given the ease of user of the editor and the character limit of 16000 for lessons, this should differentiate them sufficiently.