Updated lessons to display better. Now purely HTML. Using Markdown was interfering with some HTML markup.
Lesson bodies on lessons and lesson creator/editor no longer use Markdown.

Previously this was used to prevent Angular attempting to sanitizing the input, which would remove images.

Now the innerHTML of a div is set. The HTML is sanitized as before using DOMPurify. Angular is then explicitly told to not attempt to sanitize the input.

Added some CSS to styles.css to style codeblocks. The editor exports these with classes instead of style attributes. I was unable to find anyway to change this.