Updated group home member/non-member lists...
They now correctly add the user to the other list if their id is greater than the last user in the list.

For example, if the member list ends with user id 5 and a user with id 10 is added to the group, the user with id 10 will be added to the end of the list.
Previously this did not happen correctly.

Was due to the checks.
- If the id of the added user is less than the id of the last member in the list, they'd be added.
- If the id of the added user was greater than the id of the last member in the list, they wouldn't be added since they'd be fetched by a future API request.

However, it did not check the end of content variable. If this was true then the client would never attempt to get more members from the API.
So now, if the end of content variable is true, the user is added to the end of the list.