Fixed bug where pressing back button while a modal was open caused the site to lock up.
Site would lock up when using browser back button while a modal was open. Was because back button would send user to another page without the modal, however the jQuery (provided by bootstrap) which handled the modal openning/closing did not run the closing even.

This resulted in the background remaining dark. The modal is not longer shown. The modal is not closed by clicking the background. This made it impossible to do anything until the app was refreshed.

Fixed by adding a global jQuery event. When the browser history API has a popstate event, all open modals are manually closed.